Drive the test suite from scenario-tester
This completes the refactor away from suite-setup-util to using scenario-tester to drive the tests.
