+++ packages/util/package.json
@@ -68,7 +68,6 @@
     "ember-source": "~3.26",
     "ember-source-channel-url": "^3.0.0",
     "ember-template-lint": "^2.18.1",
-    "ember-try": "^1.4.0",
     "eslint": "^7.20.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-ember": "^10.2.0",
