--- test-packages/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  testMatch: ['<rootDir>/dynamic_suites/*.test.js'],
-};
