+++ test-packages/sample-transforms/package.json
@@ -44,7 +44,6 @@
     "ember-source": "~3.26",
     "ember-source-channel-url": "^1.1.0",
     "ember-template-lint": "^3.6.0",
-    "ember-try": "^1.0.0",
     "eslint-plugin-ember": "^7.0.0",
     "eslint-plugin-node": "^9.0.1",
     "loader.js": "^4.7.0",
