+++ test-packages/sample-transforms/testem.js
@@ -1,2 +1,2 @@
-const { testemConfig } = require('@embroider/test-support/suite-setup-util');
+const { testemConfig } = require('@embroider/test-support/testem-config');
 module.exports = testemConfig();
