+++ test-packages/support/testem-config.ts
@@ -1,0 +1,24 @@
+export function testemConfig() {
+  return {
+    test_page: 'tests/index.html?hidepassed',
+    disable_watching: true,
+    launch_in_ci: ['Chrome'],
+    launch_in_dev: ['Chrome'],
+    browser_start_timeout: 90,
+    browser_args: {
+      Chrome: {
+        ci: [
+          // --no-sandbox is needed when running Chrome inside a container
+          process.env.CI ? '--no-sandbox' : null,
+          '--headless',
+          '--disable-dev-shm-usage',
+          '--disable-software-rasterizer',
+          '--mute-audio',
+          '--remote-debugging-port=0',
+          '--window-size=1440,900',
+          `--crash-dumps-dir=${process.env.TMPDIR}`,
+        ].filter(Boolean),
+      },
+    },
+  };
+}
